+++ job-engine/api/src/main/java/org/eclipse/kapua/job/engine/exception/CleanJobDataException.java
@@ -10,17 +10,12 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.job.engine.jbatch.exception;
+package org.eclipse.kapua.job.engine.exception;
 
 import org.eclipse.kapua.model.id.KapuaId;
 
-public class JobStopppingException extends JobEngineException {
-
-    public JobStopppingException(Throwable t, KapuaId scopeId, KapuaId jobId) {
-        this(t, scopeId, jobId, null);
-    }
-
-    public JobStopppingException(Throwable t, KapuaId scopeId, KapuaId jobId, KapuaId jobExecutionId) {
-        super(KapuaJobEngineErrorCodes.JOB_STOPPING, t, scopeId, jobId, jobExecutionId);
+public class CleanJobDataException extends JobEngineException {
+    public CleanJobDataException(Throwable t, KapuaId scopeId, KapuaId jobId) {
+        super(KapuaJobEngineErrorCodes.CANNOT_CLEANUP_JOB_DATA, t, scopeId, jobId);
     }
 }
