+++ job-engine/api/src/main/java/org/eclipse/kapua/job/engine/exception/JobInvalidTargetException.java
@@ -10,12 +10,23 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.job.engine.jbatch.exception;
+package org.eclipse.kapua.job.engine.exception;
 
 import org.eclipse.kapua.model.id.KapuaId;
 
-public class CleanJobDataException extends JobEngineException {
-    public CleanJobDataException(Throwable t, KapuaId scopeId, KapuaId jobId) {
-        super(KapuaJobEngineErrorCodes.CANNOT_CLEANUP_JOB_DATA, t, scopeId, jobId);
+import java.util.Set;
+
+public class JobInvalidTargetException extends JobEngineException {
+
+    private final Set<KapuaId> targetSublist;
+
+    public JobInvalidTargetException(KapuaId scopeId, KapuaId jobId, Set<KapuaId> targetSublist) {
+        super(KapuaJobEngineErrorCodes.JOB_TARGET_INVALID, scopeId, jobId, targetSublist.toArray());
+        this.targetSublist = targetSublist;
     }
+
+    public Set<KapuaId> getTargetSublist() {
+        return targetSublist;
+    }
+
 }
