+++ job-engine/api/src/main/java/org/eclipse/kapua/job/engine/exception/KapuaJobEngineException.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.job.engine.jbatch.exception;
+package org.eclipse.kapua.job.engine.exception;
 
 import org.eclipse.kapua.KapuaException;
