+++ job-engine/jbatch/src/main/java/org/eclipse/kapua/job/engine/jbatch/JobEngineServiceJbatch.java
@@ -19,16 +19,16 @@ import org.eclipse.kapua.commons.util.ArgumentValidator;
 import org.eclipse.kapua.job.engine.JobEngineService;
 import org.eclipse.kapua.job.engine.JobStartOptions;
 import org.eclipse.kapua.job.engine.jbatch.driver.JbatchDriver;
-import org.eclipse.kapua.job.engine.jbatch.exception.CleanJobDataException;
-import org.eclipse.kapua.job.engine.jbatch.exception.JobCheckRunningException;
-import org.eclipse.kapua.job.engine.jbatch.exception.JobInvalidTargetException;
-import org.eclipse.kapua.job.engine.jbatch.exception.JobMissingStepException;
-import org.eclipse.kapua.job.engine.jbatch.exception.JobMissingTargetException;
-import org.eclipse.kapua.job.engine.jbatch.exception.JobNotRunningException;
-import org.eclipse.kapua.job.engine.jbatch.exception.JobResumingException;
-import org.eclipse.kapua.job.engine.jbatch.exception.JobRunningException;
-import org.eclipse.kapua.job.engine.jbatch.exception.JobStartingException;
-import org.eclipse.kapua.job.engine.jbatch.exception.JobStopppingException;
+import org.eclipse.kapua.job.engine.exception.CleanJobDataException;
+import org.eclipse.kapua.job.engine.exception.JobCheckRunningException;
+import org.eclipse.kapua.job.engine.exception.JobInvalidTargetException;
+import org.eclipse.kapua.job.engine.exception.JobMissingStepException;
+import org.eclipse.kapua.job.engine.exception.JobMissingTargetException;
+import org.eclipse.kapua.job.engine.exception.JobNotRunningException;
+import org.eclipse.kapua.job.engine.exception.JobResumingException;
+import org.eclipse.kapua.job.engine.exception.JobRunningException;
+import org.eclipse.kapua.job.engine.exception.JobStartingException;
+import org.eclipse.kapua.job.engine.exception.JobStoppingException;
 import org.eclipse.kapua.locator.KapuaLocator;
 import org.eclipse.kapua.locator.KapuaProvider;
 import org.eclipse.kapua.model.KapuaEntityAttributes;
@@ -190,7 +190,7 @@ public class JobEngineServiceJbatch implements JobEngineService {
         try {
             JbatchDriver.stopJob(scopeId, jobId, null);
         } catch (Exception e) {
-            throw new JobStopppingException(e, scopeId, jobId);
+            throw new JobStoppingException(e, scopeId, jobId);
         }
     }
 
@@ -231,7 +231,7 @@ public class JobEngineServiceJbatch implements JobEngineService {
         try {
             JbatchDriver.stopJob(scopeId, jobId, jobExecutionId);
         } catch (Exception e) {
-            throw new JobStopppingException(e, scopeId, jobId, jobExecutionId);
+            throw new JobStoppingException(e, scopeId, jobId, jobExecutionId);
         }
 
     }
