+++ job-engine/jbatch/src/main/java/org/eclipse/kapua/job/engine/jbatch/listener/KapuaJobListener.java
@@ -20,8 +20,8 @@ import org.eclipse.kapua.job.engine.commons.logger.JobLogger;
 import org.eclipse.kapua.job.engine.commons.model.JobTargetSublist;
 import org.eclipse.kapua.job.engine.commons.wrappers.JobContextWrapper;
 import org.eclipse.kapua.job.engine.jbatch.driver.JbatchDriver;
-import org.eclipse.kapua.job.engine.jbatch.exception.JobAlreadyRunningException;
-import org.eclipse.kapua.job.engine.jbatch.exception.JobExecutionEnqueuedException;
+import org.eclipse.kapua.job.engine.exception.JobAlreadyRunningException;
+import org.eclipse.kapua.job.engine.exception.JobExecutionEnqueuedException;
 import org.eclipse.kapua.job.engine.jbatch.setting.JobEngineSetting;
 import org.eclipse.kapua.job.engine.jbatch.setting.JobEngineSettingKeys;
 import org.eclipse.kapua.job.engine.queue.QueuedJobExecution;
